+++ cmd/frontend/internal/graphqlbackend/user.go
@@ -87,11 +87,7 @@ func (r *userResolver) Verified() bool {
 	return r.user.Verified
 }
 
-func (r *userResolver) SiteAdmin() bool {
-	// TODO(sqs): temp implementation until https://github.com/sourcegraph/sourcegraph/pull/8698
-	// is merged
-	return (&actor.Actor{Login: r.user.Username}).IsAdmin()
-}
+func (r *userResolver) SiteAdmin() bool { return r.user.SiteAdmin }
 
 // CreateUser creates a new native auth (non-SSO) user.
 func (*schemaResolver) CreateUser(ctx context.Context, args *struct {